Update GiftLottery.sol
Created a method/function called pseudoRandom on line 50, which will take the role of a random generator, which by a gift id load the
gift and calculate keccak256 on its title, description, URL and gift id. Someone could be left with no gifts and some could be left with two or more... The method will "randomly" generate the id of the participant to which the gift will be sent too.